+++ tests/conftest.py
@@ -447,6 +447,79 @@ def postgres_server():
         subprocess.run(["docker", "stop", "hypha-postgres"], check=True)
 
 
+@pytest_asyncio.fixture(name="pg_engine")
+async def pg_engine(postgres_server):
+    """Fixture to set up PostgreSQL async engine."""
+    # Import here to avoid circular dependencies
+    from sqlalchemy.ext.asyncio import create_async_engine
+    import asyncpg
+
+    # Wait for PostgreSQL to be ready
+    max_retries = 20
+    retry_delay = 1.0
+
+    for attempt in range(max_retries):
+        try:
+            # Connect directly with asyncpg to check if ready
+            conn = await asyncpg.connect(
+                host="127.0.0.1",
+                port=POSTGRES_PORT,
+                user=POSTGRES_USER,
+                password=POSTGRES_PASSWORD,
+                database=POSTGRES_DB
+            )
+            try:
+                # Create pgvector extension if not exists
+                await conn.execute("CREATE EXTENSION IF NOT EXISTS vector")
+                print("pgvector extension ready")
+            except asyncpg.exceptions.DuplicateObjectError:
+                print("pgvector extension already exists")
+            except Exception as e:
+                print(f"Warning: Could not create pgvector extension: {e}")
+            finally:
+                await conn.close()
+            break
+        except (asyncpg.exceptions.InvalidPasswordError, ConnectionRefusedError, OSError) as e:
+            if attempt < max_retries - 1:
+                print(f"Waiting for PostgreSQL to be ready (attempt {attempt + 1}/{max_retries})...")
+                await asyncio.sleep(retry_delay)
+            else:
+                raise Exception(f"Failed to connect to PostgreSQL after {max_retries} attempts: {e}")
+
+    # Now create the SQLAlchemy engine
+    engine = create_async_engine(
+        POSTGRES_URI,
+        echo=False,
+        pool_pre_ping=True,
+    )
+
+    yield engine
+
+    # Cleanup
+    await engine.dispose()
+
+
+@pytest_asyncio.fixture(name="pgvector_search_engine")
+async def pgvector_search_engine(pg_engine):
+    """Fixture to set up PgVectorSearchEngine instance."""
+    from hypha.pgvector import PgVectorSearchEngine
+
+    # Create a fresh instance for each test
+    search_engine = PgVectorSearchEngine(engine=pg_engine, prefix="test")
+    yield search_engine
+
+    # Cleanup: Drop all collections created in test
+    try:
+        collections = await search_engine.list_collections()
+        for collection in collections:
+            try:
+                await search_engine.delete_collection(collection["name"])
+            except Exception as e:
+                print(f"Warning: Could not delete collection {collection['name']}: {e}")
+    except Exception as e:
+        print(f"Warning during cleanup: {e}")
+
+
 @pytest_asyncio.fixture(name="redis_server", scope="session")
 def redis_server():
     """Start a redis server as test fixture and tear down after test."""
